reference/advanced-autodistill/main-commit/requirements.txt
opencv-python>=4.7.0
matplotlib>=3.7.0
pyyaml>=6.0
wheel>=0.40.0
autodistill>=0.1.9
autodistill-grounded-sam-2>=0.1.0
autodistill-yolov8>=0.1.0
autodistill-florence-2>=0.1.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic_settings>=2.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.24.1
python-dotenv>=1.0.0
pytest>=7.3.1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
seaborn>=0.12.0
gradio>=3.40.0
clip @ git+https://github.com/openai/CLIP.git